Move hyperdrive.resource.relative_path to fileshelper
